Struct rusoto_cur::CostAndUsageReportClient

pub struct CostAndUsageReportClient { /* fields omitted */ }

A client for the AWS Cost and Usage Report Service API.

Methods

impl CostAndUsageReportClient

pub fn new(region: Region) -> CostAndUsageReportClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CostAndUsageReportClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,