Trait rusoto_cur::CostAndUsageReport

pub trait CostAndUsageReport {
    fn delete_report_definition(
        &self, 
        input: &DeleteReportDefinitionRequest
    ) -> Result<DeleteReportDefinitionResponse, DeleteReportDefinitionError>;
    fn describe_report_definitions(
        &self, 
        input: &DescribeReportDefinitionsRequest
    ) -> Result<DescribeReportDefinitionsResponse, DescribeReportDefinitionsError>;
    fn put_report_definition(
        &self, 
        input: &PutReportDefinitionRequest
    ) -> Result<PutReportDefinitionResponse, PutReportDefinitionError>;
}

Trait representing the capabilities of the AWS Cost and Usage Report Service API. AWS Cost and Usage Report Service clients implement this trait.

Required Methods

fn delete_report_definition(
&self,
input: &DeleteReportDefinitionRequest
) -> Result<DeleteReportDefinitionResponse, DeleteReportDefinitionError>

Delete a specified report definition

fn describe_report_definitions(
&self,
input: &DescribeReportDefinitionsRequest
) -> Result<DescribeReportDefinitionsResponse, DescribeReportDefinitionsError>

Describe a list of report definitions owned by the account

fn put_report_definition(
&self,
input: &PutReportDefinitionRequest
) -> Result<PutReportDefinitionResponse, PutReportDefinitionError>

Create a new report definition