Trait rusoto_cur::CostAndUsageReport

pub trait CostAndUsageReport {
    fn delete_report_definition(
        &self, 
        input: DeleteReportDefinitionRequest
    ) -> RusotoFuture<DeleteReportDefinitionResponse, DeleteReportDefinitionError>;

    fn describe_report_definitions(
        &self, 
        input: DescribeReportDefinitionsRequest
    ) -> RusotoFuture<DescribeReportDefinitionsResponse, DescribeReportDefinitionsError>;

    fn put_report_definition(
        &self, 
        input: PutReportDefinitionRequest
    ) -> RusotoFuture<PutReportDefinitionResponse, PutReportDefinitionError>;
}

Trait representing the capabilities of the AWS Cost and Usage Report Service API. AWS Cost and Usage Report Service clients implement this trait.

Required methods

fn delete_report_definition(
&self,
input: DeleteReportDefinitionRequest
) -> RusotoFuture<DeleteReportDefinitionResponse, DeleteReportDefinitionError>

Deletes the specified report.

fn describe_report_definitions(
&self,
input: DescribeReportDefinitionsRequest
) -> RusotoFuture<DescribeReportDefinitionsResponse, DescribeReportDefinitionsError>

Lists the AWS Cost and Usage reports available to this account.

fn put_report_definition(
&self,
input: PutReportDefinitionRequest
) -> RusotoFuture<PutReportDefinitionResponse, PutReportDefinitionError>

Creates a new report using the description that you provide.

Loading content...